Cmake on Windows: set ouput path
This forces all output files (libraries and executables) to one directory.
Then debugging the demo application is possible. Just the mnobd.cdc has to
be copied to the very same directory. The INSTALL target does not copy the
pdb files to the installation directory, but those contain the debug
symbols and are required for debugging.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>